Extract model breakdown helpers to satisfy type body length
